requirements.txt
astronomer-cosmos[dbt.all]
apache-airflow-providers-postgres
pymysql
pandas
boto3
sqlalchemy
clickhouse_driver
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy